requirements.txt
aiohttp==3.6.2
bs4==0.0.1
discord.py==1.3.4
dnspython==2.0.0
git+https://github.com/WhyNoLetty/Translation.git
motor==2.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
uptime==3.0.1
